Button placement
Moved button from above the map to the corner of it
